Clean up a lot of irrelevant code and simplify structure
