--- source/dotnet/subsystem-tests/ExternalProjects/CalculationApplication/Model/Calculations/CalculationExecutionState.cs
@@ -1,12 +1,0 @@
-﻿namespace Energinet.DataHub.Wholesale.Calculations.Application.Model.Calculations;
-
-/// <summary>
-/// IMPORTANT: Do not change numeric values as it'll affect persistence or communication.
-/// </summary>
-public enum CalculationExecutionState
-{
-    /// <summary>
-    /// The calculation is created.
-    /// </summary>
-    Created = -2,
-}
