--- source/dotnet/subsystem-tests/ExternalProjects/CalculationApplication/Model/Calculations/CalculationJobId.cs
@@ -1,3 +1,0 @@
-﻿namespace Energinet.DataHub.Wholesale.Calculations.Application.Model.Calculations;
-
-public sealed record CalculationJobId(long Id);
