--- source/dotnet/subsystem-tests/ExternalProjects/CalculationInterfaces/BusinessValidationException.cs
@@ -1,9 +1,0 @@
-﻿namespace Energinet.DataHub.Wholesale.Calculations.Interfaces;
-
-public class BusinessValidationException : Exception
-{
-    public BusinessValidationException(string message)
-        : base(message)
-    {
-    }
-}
