--- source/dotnet/subsystem-tests/ExternalProjects/CalculationResultsInfrastructure/DeltaTableCalculationType.cs
@@ -1,11 +1,0 @@
-﻿namespace Energinet.DataHub.Wholesale.CalculationResults.Infrastructure.SqlStatements.DeltaTableConstants;
-
-public static class DeltaTableCalculationType
-{
-    public const string Aggregation = "aggregation";
-    public const string BalanceFixing = "balance_fixing";
-    public const string WholesaleFixing = "wholesale_fixing";
-    public const string FirstCorrectionSettlement = "first_correction_settlement";
-    public const string SecondCorrectionSettlement = "second_correction_settlement";
-    public const string ThirdCorrectionSettlement = "third_correction_settlement";
-}
