+++ source/dotnet/subsystem-tests/Features/Calculations/Fixtures/CalculationJobScenarioFixture.cs
@@ -7,11 +7,9 @@ using Energinet.DataHub.Core.Databricks.SqlStatementExecution;
 using Energinet.DataHub.Core.TestCommon;
 using Energinet.DataHub.Core.TestCommon.Xunit.Extensions;
 using Energinet.DataHub.Core.TestCommon.Xunit.LazyFixture;
-using Energinet.DataHub.Wholesale.Calculations.Application.Model;
-using Energinet.DataHub.Wholesale.Calculations.Application.Model.Calculations;
-using Energinet.DataHub.Wholesale.Calculations.Infrastructure.Calculations;
 using Energinet.DataHub.Wholesale.SubsystemTests.Features.Calculations.States;
 using Energinet.DataHub.Wholesale.SubsystemTests.Fixtures;
+using Energinet.DataHub.Wholesale.SubsystemTests.Models;
 using Microsoft.Azure.Databricks.Client;
 using Microsoft.Azure.Databricks.Client.Models;
 using Microsoft.Extensions.Configuration;
