+++ source/dotnet/subsystem-tests/Features/Calculations/WholesaleFixingCalculationJobScenario.cs
@@ -2,9 +2,8 @@
 using Energinet.DataHub.Core.TestCommon.Xunit.Attributes;
 using Energinet.DataHub.Core.TestCommon.Xunit.LazyFixture;
 using Energinet.DataHub.Core.TestCommon.Xunit.Orderers;
-using Energinet.DataHub.Wholesale.Calculations.Application.Model;
-using Energinet.DataHub.Wholesale.Calculations.Application.Model.Calculations;
 using Energinet.DataHub.Wholesale.SubsystemTests.Features.Calculations.Fixtures;
+using Energinet.DataHub.Wholesale.SubsystemTests.Models;
 using FluentAssertions;
 using FluentAssertions.Execution;
 using NodaTime;
@@ -41,15 +40,11 @@ public class WholesaleFixingCalculationJobScenario : SubsystemTestsBase<Calculat
         var createdTime = SystemClock.Instance.GetCurrentInstant();
         var createdByUserId = Guid.Parse("DED7734B-DD56-43AD-9EE8-0D7EFDA6C783");
         Fixture.ScenarioState.CalculationJobInput = new Calculation(
-            createdTime: createdTime,
-            calculationType: Common.Interfaces.Models.CalculationType.WholesaleFixing,
+            calculationType: CalculationType.WholesaleFixing,
             gridAreaCodes: new List<GridAreaCode> { new("804") },
             periodStart: Instant.FromDateTimeOffset(new DateTimeOffset(2023, 1, 31, 23, 0, 0, TimeSpan.Zero)),
             periodEnd: Instant.FromDateTimeOffset(new DateTimeOffset(2023, 2, 28, 23, 0, 0, TimeSpan.Zero)),
-            scheduledAt: createdTime, // Schedule to run immediately
-            dateTimeZone: DateTimeZoneProviders.Tzdb.GetZoneOrNull("Europe/Copenhagen")!,
             createdByUserId: createdByUserId,
-            version: createdTime.ToDateTimeUtc().Ticks,
             false);
     }
