+++ source/dotnet/subsystem-tests/Models/CalculationJobId.cs
@@ -1,0 +1,3 @@
+﻿namespace Energinet.DataHub.Wholesale.SubsystemTests.Models;
+
+public sealed record CalculationJobId(long Id);
