+++ source/dotnet/subsystem-tests/Models/GridAreaCode.cs
@@ -1,0 +1,3 @@
+﻿namespace Energinet.DataHub.Wholesale.SubsystemTests.Models;
+
+public sealed record GridAreaCode(string Code);
